backend/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
greenlet==3.3.2
gunicorn==25.3.0
idna==3.11
packaging==26.2
pillow==12.2.0
playwright==1.58.0
psycopg2-binary==2.9.12
pyee==13.0.1
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.0
resend==2.29.0
schedule==1.2.2
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
